backend/requirements.txt
# Python 3.11+ (PyCaret requires <=3.11)

# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Data Storage
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
pyarrow>=16.0.0

# DuckDB (已废弃，保留用于数据迁移)
# duckdb>=1.4.0

# Data Sources
tushare>=1.4.19
akshare>=1.16.72

# Data Processing
polars>=1.0.0
pandas>=2.1.0,<2.3.0
numpy>=1.26.0,<2.1.0

# Quantitative
vectorbt>=0.26.0

# Machine Learning
pycaret>=3.3.0
optuna>=3.6.0
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Task Scheduling
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
